CI: use Parallel-Lint
Parallel-Lint, as the name implies, runs PHP lint in parallel, making it faster than native linting.
It also provides more helpful output with a syntax highlighter code snippet showing the context of syntax errors.

As we're now doing a full `composer install` on every build anyway, we may as well use this tool.